Undo the commit moving the files; will do that in separate PR
