+++ torchchat.py
@@ -87,7 +87,7 @@ if __name__ == "__main__":
         eval_main(args)
     elif args.command == "export":
         check_args(args, "export")
-        from torchchat.export import main as export_main
+        from export import main as export_main
 
         export_main(args)
     elif args.command == "download":
